refactor: organize amazon integration containers
