+++ src/core/integrations/integrations/integrations-show/containers/properties/containers/amazon-properties/configs.ts
@@ -1,10 +1,10 @@
-import { FieldType } from "../../../../../../shared/utils/constants";
-import { amazonPropertiesQuery, getAmazonPropertyQuery } from "../../../../../../shared/api/queries/salesChannels.js";
-import { propertiesQuery } from "../../../../../../shared/api/queries/properties.js";
-import { updateAmazonPropertyMutation } from "../../../../../../shared/api/mutations/salesChannels.js";
-import { ListingConfig } from "../../../../../../shared/components/organisms/general-listing/listingConfig";
-import { FormConfig, FormType } from '../../../../../../shared/components/organisms/general-form/formConfig';
-import { SearchConfig } from "../../../../../../shared/components/organisms/general-search/searchConfig";
+import { FieldType } from "../../../../../../../../shared/utils/constants";
+import { amazonPropertiesQuery, getAmazonPropertyQuery } from "../../../../../../../../shared/api/queries/salesChannels.js";
+import { propertiesQuery } from "../../../../../../../../shared/api/queries/properties.js";
+import { updateAmazonPropertyMutation } from "../../../../../../../../shared/api/mutations/salesChannels.js";
+import { ListingConfig } from "../../../../../../../../shared/components/organisms/general-listing/listingConfig";
+import { FormConfig, FormType } from '../../../../../../../../shared/components/organisms/general-form/formConfig';
+import { SearchConfig } from "../../../../../../../../shared/components/organisms/general-search/searchConfig";
 
 export const amazonPropertyEditFormConfigConstructor = (
   t: Function,
