+++ src/core/integrations/integrations/integrations-show/containers/properties/containers/amazon-properties/index.ts
