+++ src/core/integrations/integrations/integrations-show/containers/rules/containers/amazon-product-types/index.ts
